requirements.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.2.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.6.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
idna==2.9
importlib-metadata==1.6.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
more-itertools==8.2.0
packaging==20.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.8.1
pycodestyle==2.6.0
Pygments==2.7.4
PyJWT==2.0.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytils==0.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==5.0.5
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
